+++ src/errors/rust_errors/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "no_std")]
+pub mod core_lib;
+
+#[cfg(not(feature = "no_std"))]
+pub mod std_lib;