+++ src/errors/tokio.rs
@@ -10,49 +10,6 @@ impl From<tokio::time::error::Error> for TheErrorType {
     }
 }
 
-impl From<tokio::io::Error> for TheErrorType {
-    fn from(value: tokio::io::Error) -> Self {
-        Self {
-            error_type: SystemErrorCodes::FileError,
-            error_content: value.to_string()
-        }
-    }
-}
-
-impl From<tokio::io::ErrorKind> for TheErrorType {
-    fn from(value: tokio::io::ErrorKind) -> Self {
-        let error_content = value.to_string();
-        let error_type = match value {
-            tokio::io::ErrorKind::NotFound => {SystemErrorCodes::NotFound}
-            tokio::io::ErrorKind::PermissionDenied => {SystemErrorCodes::PermissionDenied}
-            tokio::io::ErrorKind::ConnectionRefused => {SystemErrorCodes::ConnectionRefused}
-            tokio::io::ErrorKind::ConnectionReset => {SystemErrorCodes::ConnectionReset}
-            tokio::io::ErrorKind::ConnectionAborted => {SystemErrorCodes::ConnectionAborted}
-            tokio::io::ErrorKind::NotConnected => {SystemErrorCodes::NotConnected}
-            tokio::io::ErrorKind::AddrInUse => {SystemErrorCodes::AddrInUse}
-            tokio::io::ErrorKind::AddrNotAvailable => {SystemErrorCodes::AddrNotAvailable}
-            tokio::io::ErrorKind::BrokenPipe => {SystemErrorCodes::BrokenPipe}
-            tokio::io::ErrorKind::AlreadyExists => {SystemErrorCodes::AlreadyExists}
-            tokio::io::ErrorKind::WouldBlock => {SystemErrorCodes::WouldBlock}
-            tokio::io::ErrorKind::InvalidInput => {SystemErrorCodes::InvalidInput}
-            tokio::io::ErrorKind::InvalidData => {SystemErrorCodes::InvalidData}
-            tokio::io::ErrorKind::TimedOut => {SystemErrorCodes::TimedOut}
-            tokio::io::ErrorKind::WriteZero => {SystemErrorCodes::WriteZero}
-            tokio::io::ErrorKind::Interrupted => {SystemErrorCodes::Interrupted}
-            tokio::io::ErrorKind::Unsupported => {SystemErrorCodes::Unsupported}
-            tokio::io::ErrorKind::UnexpectedEof => {SystemErrorCodes::UnexpectedEof}
-            tokio::io::ErrorKind::OutOfMemory => {SystemErrorCodes::OutOfMemory}
-            tokio::io::ErrorKind::Other => {SystemErrorCodes::Other}
-            _ => {SystemErrorCodes::UnstableErrorType }
-        };
-
-        Self {
-            error_type,
-            error_content
-        }
-    }
-}
-
 impl<T> From<tokio::sync::SetError<T>> for TheErrorType {
     fn from(value: tokio::sync::SetError<T>) -> Self {
         Self {
@@ -197,7 +154,7 @@ impl From<tokio::sync::watch::error::RecvError> for TheErrorType {
 impl From<tokio::task::JoinError> for TheErrorType {
     fn from(value: tokio::task::JoinError) -> Self {
         Self {
-            error_type: SystemErrorCodes::TaskError,
+            error_type: SystemErrorCodes::JoinError,
             error_content: value.to_string()
         }
     }